refactor: 사이드바 로직 분리
