+++ src/features/sign-out/lib/useSignOut.ts
@@ -1,0 +1,14 @@
+import { useRouter } from "next/navigation";
+
+export const useSignOut = () => {
+  const router = useRouter();
+
+  const handleSignout = async () => {
+    await fetch("/api/auth/signout", { method: "POST" });
+    sessionStorage.removeItem("access_token");
+    sessionStorage.removeItem("user");
+    router.push("/signin");
+  };
+
+  return { handleSignout };
+};
