+++ src/widgets/sidebar/ui/sidebarMenu.tsx
